Added rerun for fimo scan
since fimo scan is SO DAMN SLOW I now run the scan threaded by TF and once the scan is done I immediately convert it to the bed format and save it. It only saves if the scan is complete for the given tf.
The rerun flag checks the motif dir for all complete tfs, then subtracts these from the list that is used to thread the fimo scan.