Statistics module update
Fits linear regression to central mass of TFs (assuming 25% outliers)-- uses sklearn.covariance EllipticEnvelope to define central 75%. 
Then it uses the normal distribution of the residuals of 75% of the TFs to assign a pval to each TF.
As number of outliers increases so do the number of significant (pval <0.05) TFs due to using the central mass residuals to calculate sigma and mu (sigma decreases).
25% outlier was selected as the number of significant TFs seems pretty stable from 20-30% outliers.
Plots gc content, rbg plot significance and the ellip. plot used to create the regression.